requirements.txt
python-telegram-bot==21.10
groq==0.18.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
